REL: Release 1.1.0 (#342)
- Bump version: 1.0.1 → 1.1.0
- Updates AUTHORS.md

### Major Updates and Feature Improvements
- Adds DICOM support.
- Updates license to more permissive BSD 3-clause.
- Adds [MicroNet](https://doi.org/10.1016/j.media.2018.12.003) model.
- Improves support for `tiff` files.
  - Adds a check for tiles in a TIFF file when opening.
  - Uses OpenSlide to read a TIFF if it has tiles instead of OpenCV (VirtualWSIReader).
  - Adds a fallback to tifffile if it is tiled but openslide cannot read it
  (e.g. jp2k or jpegxl tiles).
- Adds support for multi-channel images (HxWxC).
- Fixes performance issues in `semantic_segmentor.py`.
  - Performance gain measurement: 21.67s (new) vs 45.564 (old) using a 4k x 4k WSI.
  - External Contribution from @ByteHexler.
- Adds benchmark for Annotations Store.

### Changes to API
- None

### Bug Fixes and Other Changes
- Enhances the error messages to be more informative.
- Fixes Flake8 Errors, typos.
  - Fixes patch predictor models based after fixing a typo.
- Bug fixes in Graph functions.
- Adds documentation for docker support.
- General tidying up of docstrings.
- Adds metrics to readthedocs/docstrings for pretrained models.

### Development related changes
- Adds `pydicom` and `wsidicom` as dependency.
- Updates dependencies.
- Fixes Travis detection and makes improvements to run tests faster on Travis.
- Adds Dependabot to automatically update dependencies.
- Improves CLI definitions to make it easier to integrate new functions.
- Fixes compile options for test_annotation_stores.py

+++ tiatoolbox/utils/misc.py
@@ -1,23 +1,3 @@
-# ***** BEGIN GPL LICENSE BLOCK *****
-#
-# This program is free software; you can redistribute it and/or
-# modify it under the terms of the GNU General Public License
-# as published by the Free Software Foundation; either version 2
-# of the License, or (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program; if not, write to the Free Software Foundation,
-# Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
-#
-# The Original Code is Copyright (C) 2021, TIA Centre, University of Warwick
-# All rights reserved.
-# ***** END GPL LICENSE BLOCK *****
-
 """Miscellaneous small functions repeatedly used in tiatoolbox."""
 import copy
 import json
@@ -41,10 +21,12 @@ def split_path_name_ext(full_path):
     """Split path of a file to directory path, file name and extensions.
 
     Args:
-        full_path (str or pathlib.Path): Path to a file
+        full_path (str or pathlib.Path):
+            Path to a file.
 
     Returns:
-        tuple: Three parts of the input file path:
+        tuple:
+            Three parts of the input file path:
             - :py:obj:`pathlib.Path` - Parent directory path
             - :py:obj:`str` - File name
             - :py:obj:`list(str)` - File extensions
@@ -63,13 +45,16 @@ def grab_files_from_dir(input_path, file_types=("*.jpg", "*.png", "*.tif")):
     """Grab file paths specified by file extensions.
 
     Args:
-        input_path (str or pathlib.Path): Path to the directory where files
+        input_path (str or pathlib.Path):
+            Path to the directory where files
             need to be searched.
-        file_types (str or tuple(str)): File types (extensions) to be searched.
+        file_types (str or tuple(str)):
+            File types (extensions) to be searched.
 
     Returns:
-        list: File paths as a python list. It has been sorted to ensure
-            same ordering across platforms.
+        list:
+            File paths as a python list. It has been sorted to ensure
+            the same ordering across platforms.
 
     Examples:
         >>> from tiatoolbox import utils
@@ -94,12 +79,25 @@ def grab_files_from_dir(input_path, file_types=("*.jpg", "*.png", "*.tif")):
     return list(files_grabbed)
 
 
-def save_yaml(input_dict, output_path="output.yaml"):
+def save_yaml(
+    input_dict: dict,
+    output_path="output.yaml",
+    parents: bool = False,
+    exist_ok: bool = False,
+):
     """Save dictionary as yaml.
 
     Args:
-        input_dict (dict): A variable of type 'dict'
-        output_path (str or pathlib.Path): Path to save the output file
+        input_dict (dict):
+            A variable of type 'dict'.
+        output_path (str or pathlib.Path):
+            Path to save the output file.
+        parents (bool):
+            Make parent directories if they do not exist. Default is
+            False.
+        exist_ok (bool):
+            Overwrite the output file if it exists. Default is False.
+
 
     Returns:
 
@@ -109,19 +107,25 @@ def save_yaml(input_dict, output_path="output.yaml"):
         >>> utils.misc.save_yaml(input_dict, './hello.yaml')
 
     """
-    with open(str(pathlib.Path(output_path)), "w") as yaml_file:
+    path = pathlib.Path(output_path)
+    if path.exists() and not exist_ok:
+        raise FileExistsError("File already exists.")
+    if parents:
+        path.parent.mkdir(parents=True, exist_ok=True)
+    with open(  # skipcq: PTC-W6004: PTC-W6004
+        str(pathlib.Path(output_path)), "w"
+    ) as yaml_file:
         yaml.dump(input_dict, yaml_file)
 
 
-def imwrite(image_path, img):
+def imwrite(image_path, img) -> None:
     """Write numpy array to an image.
 
     Args:
-        image_path (str or pathlib.Path): file path (including extension)
-            to save image
-        img (:class:`numpy.ndarray`): image array of dtype uint8, MxNx3
-
-    Returns:
+        image_path (str or pathlib.Path):
+            File path (including extension) to save image to.
+        img (:class:`numpy.ndarray`):
+            Image array of dtype uint8, MxNx3.
 
     Examples:
         >>> from tiatoolbox import utils
@@ -139,11 +143,14 @@ def imread(image_path, as_uint8=True):
     """Read an image as numpy array.
 
     Args:
-        image_path (str or pathlib.Path): File path (including extension) to read image.
-        as_uint8 (bool): Read an image in uint8 format.
+        image_path (str or pathlib.Path):
+            File path (including extension) to read image.
+        as_uint8 (bool):
+            Read an image in uint8 format.
 
     Returns:
-        img (:class:`numpy.ndarray`): Image array of dtype uint8, MxNx3.
+        :class:`numpy.ndarray`:
+            Image array of dtype uint8, MxNx3.
 
     Examples:
         >>> from tiatoolbox import utils
@@ -168,12 +175,14 @@ def load_stain_matrix(stain_matrix_input):
     """Load a stain matrix as a numpy array.
 
     Args:
-        stain_matrix_input (ndarray or str, pathlib.Path): either a 2x3 / 3x3
-            numpy array or a path to a saved .npy / .csv file. If using a .csv file,
-            there should be no column headers provided
+        stain_matrix_input (ndarray or str, pathlib.Path):
+            Either a 2x3 or 3x3 numpy array or a path to a saved .npy /
+            .csv file. If using a .csv file, there should be no column
+            headers provided
 
     Returns:
-        stain_matrix (:class:`numpy.ndarray`): the loaded stain matrix.
+        stain_matrix (:class:`numpy.ndarray`):
+            The loaded stain matrix.
 
     Examples:
         >>> from tiatoolbox import utils
@@ -191,8 +200,8 @@ def load_stain_matrix(stain_matrix_input):
         if suffixes[-1] == ".csv":
             return pd.read_csv(stain_matrix_input).to_numpy()
 
-        if suffixes[-1] == ".npy":
-            return np.load(str(stain_matrix_input))
+        # only other option left for suffix[-1] is .npy
+        return np.load(str(stain_matrix_input))
 
     if isinstance(stain_matrix_input, np.ndarray):
         return stain_matrix_input
@@ -206,11 +215,14 @@ def get_luminosity_tissue_mask(img, threshold):
     """Get tissue mask based on the luminosity of the input image.
 
     Args:
-        img (:class:`numpy.ndarray`): input image used to obtain tissue mask.
-        threshold (float): luminosity threshold used to determine tissue area.
+        img (:class:`numpy.ndarray`):
+            Input image used to obtain tissue mask.
+        threshold (float):
+            Luminosity threshold used to determine tissue area.
 
     Returns:
-        tissue_mask (:class:`numpy.ndarray`): binary tissue mask.
+        tissue_mask (:class:`numpy.ndarray`):
+            Binary tissue mask.
 
     Examples:
         >>> from tiatoolbox import utils
@@ -225,7 +237,7 @@ def get_luminosity_tissue_mask(img, threshold):
 
     # check it's not empty
     if tissue_mask.sum() == 0:
-        raise ValueError("Empty tissue mask computed")
+        raise ValueError("Empty tissue mask computed.")
 
     return tissue_mask
 
@@ -245,7 +257,8 @@ def mpp2common_objective_power(
             (1, 1.25, 2, 2.5, 4, 5, 10, 20, 40, 60, 90, 100).
 
     Returns:
-        float: Objective power approximation.
+        float:
+            Objective power approximation.
 
     Examples:
         >>> mpp2common_objective_power(0.253)
@@ -282,7 +295,8 @@ def objective_power2mpp(objective_power):
         objective_power (float or tuple(float)): Objective power.
 
     Returns:
-        numpy.ndarray: Microns per-pixel (MPP) approximations.
+        :class:`numpy.ndarray`:
+            Microns per-pixel (MPP) approximations.
 
     Examples:
         >>> objective_power2mpp(40)
@@ -306,7 +320,8 @@ def mpp2objective_power(mpp):
         mpp (float or tuple(float)): Microns per-pixel.
 
     Returns:
-        :class:`numpy.ndarray`: Objective power approximations.
+        :class:`numpy.ndarray`:
+            Objective power approximations.
 
     Examples:
         >>> mpp2objective_power(0.25)
@@ -334,7 +349,8 @@ def contrast_enhancer(img, low_p=2, high_p=98):
             high_p should always be greater than low_p.
 
     Returns:
-        img (:class:`numpy.ndarray`): Image (uint8) with contrast enhanced.
+        img (:class:`numpy.ndarray`):
+            Image (uint8) with contrast enhanced.
 
     Raises:
         AssertionError: Internal errors due to invalid img type.
@@ -380,7 +396,7 @@ def __numpy_array_to_table(input_table):
     if input_table.shape[1] == 3:
         return pd.DataFrame(input_table, columns=["x", "y", "class"])
 
-    raise ValueError("numpy table should be of format `x, y` or " "`x, y, class`")
+    raise ValueError("Numpy table should be of format `x, y` or `x, y, class`.")
 
 
 def __assign_unknown_class(input_table):
@@ -393,7 +409,8 @@ def __assign_unknown_class(input_table):
         table (:class:`pd.DataFrame`): Pandas DataFrame with desired features.
 
     Raises:
-        ValueError: If the number of columns is not equal to 2 or 3.
+        ValueError:
+            If the number of columns is not equal to 2 or 3.
 
     """
     if input_table.shape[1] not in [2, 3]:
@@ -421,7 +438,8 @@ def read_locations(input_table):
         pd.DataFrame: DataFrame with x, y location and class type.
 
     Raises:
-        FileNotSupported: If the path to input table is not of supported type.
+        FileNotSupported:
+            If the path to input table is not of supported type.
 
     Examples:
         >>> from tiatoolbox.utils.misc import read_locations
@@ -486,7 +504,8 @@ def conv_out_size(in_size, kernel_size=1, padding=0, stride=1):
         stride (int): Stride size.
 
     Returns:
-        int: Output size / number of features.
+        int:
+            Output size / number of features.
 
     Examples:
         >>> from tiatoolbox import utils
@@ -520,14 +539,17 @@ def parse_cv2_interpolaton(interpolation: Union[str, int]) -> int:
         https://docs.opencv.org/4.0.0/da/d54/group__imgproc__transform.html#ga5bb5a1fea74ea38e1a5445ca803ff121
 
     Args:
-        interpolation (Union[str, int]): Interpolation mode string.
-            Possible values are: nearest, linear, cubic, lanczos, area.
+        interpolation (Union[str, int]):
+            Interpolation mode string. Possible values are: nearest,
+            linear, cubic, lanczos, area.
 
     Raises:
-        ValueError: Invalid interpolation mode.
+        ValueError:
+            Invalid interpolation mode.
 
     Returns:
-        int: OpenCV (cv2) interpolation enum.
+        int:
+            OpenCV (cv2) interpolation enum.
 
     """
     if isinstance(interpolation, str):
@@ -549,11 +571,14 @@ def assert_dtype_int(input_var, message="Input must be integer."):
     """Generate error if dtype is not int.
 
     Args:
-        input_var (ndarray): input variable to be tested.
-        message (str): Error message to be displayed.
+        input_var (ndarray):
+            Input variable to be tested.
+        message (str):
+            Error message to be displayed.
 
-    Returns:
-        Generates an AssertionError message if input is not an int.
+    Raises:
+        AssertionError:
+            If input_var is not of type int.
 
     """
     if not np.issubdtype(np.array(input_var).dtype, np.integer):
@@ -660,7 +685,12 @@ def __walk_dict(dct):
         dct[k] = __walk_list_dict(v)
 
 
-def save_as_json(data, save_path):
+def save_as_json(
+    data: Union[dict, list],
+    save_path: Union[str, pathlib.Path],
+    parents: bool = False,
+    exist_ok: bool = False,
+):
     """Save data to a json file.
 
     The function will deepcopy the `data` and then jsonify the content
@@ -668,31 +698,45 @@ def save_as_json(data, save_path):
     `bool` and their np.ndarray respectively.
 
     Args:
-        data (dict or list): Input data to save.
-        save_path (str): Output to save the json of `input`.
+        data (dict or list):
+            Input data to save.
+        save_path (str):
+            Output to save the json of `input`.
+        parents (bool):
+            Make parent directories if they do not exist. Default is
+            False.
+        exist_ok (bool):
+            Overwrite the output file if it exists. Default is False.
+
 
     """
     shadow_data = copy.deepcopy(data)  # make a copy of source input
     if not isinstance(shadow_data, (dict, list)):
-        raise ValueError(f"`data` type {type(data)} is not [dict, list].")
+        raise ValueError(f"Type of `data` ({type(data)}) must be in (dict, list).")
 
     if isinstance(shadow_data, dict):
         __walk_dict(shadow_data)
     else:
         __walk_list(shadow_data)
 
-    with open(save_path, "w") as handle:
+    save_path = pathlib.Path(save_path)
+    if save_path.exists() and not exist_ok:
+        raise FileExistsError("File already exists.")
+    if parents:
+        save_path.parent.mkdir(parents=True, exist_ok=True)
+    with open(save_path, "w") as handle:  # skipcq: PTC-W6004
         json.dump(shadow_data, handle)
 
 
-def select_device(on_gpu):
+def select_device(on_gpu: bool) -> str:
     """Selects the appropriate device as requested.
 
     Args:
         on_gpu (bool): Selects gpu if True.
 
     Returns:
-        device (str): "gpu" if on_gpu is True otherwise returns "cpu"
+        str:
+            "gpu" if on_gpu is True otherwise returns "cpu"
 
     """
     if on_gpu:
@@ -709,7 +753,8 @@ def model_to(on_gpu, model):
         model (torch.nn.Module): PyTorch defined model.
 
     Returns:
-        model (torch.nn.Module):
+        torch.nn.Module:
+            The model after being moved to cpu/gpu.
 
     """
     if on_gpu:  # DataParallel work only for cuda
@@ -722,16 +767,17 @@ def model_to(on_gpu, model):
 def get_bounding_box(img):
     """Get bounding box coordinate information.
 
-    Given an image with zero and non-zero values. This function
-    will return the the minimal box that contains all non-zero
-    values.
+    Given an image with zero and non-zero values. This function will
+    return the minimal box that contains all non-zero values.
 
     Args:
-        img (ndarray): Image to get the bounding box.
+        img (ndarray):
+            Image to get the bounding box.
 
     Returns:
-        bound (ndarray): Coordinates of the box in the form of
-            [start_x, start_y, end_x, end_y].
+        bound (ndarray):
+            Coordinates of the box in the form of `[start_x, start_y,
+            end_x, end_y]`.
 
     """
     rows = np.any(img, axis=1)
@@ -749,106 +795,41 @@ def string_to_tuple(in_str):
     """Splits input string to tuple at ','.
 
     Args:
-        in_str (str): input string.
+        in_str (str):
+            input string.
 
     Returns:
-        tuple (tuple of str): Returns a tuple of strings by splitting in_str at ','.
+        tuple:
+            Returns a tuple of strings by splitting in_str at ','.
 
     """
     return tuple(substring.strip() for substring in in_str.split(","))
 
 
-def prepare_file_dir_cli(img_input, output_path, file_types, mode, sub_dirname):
-    """Prepares CLI for running code on multiple files or a directory.
-
-    Checks for existing directories to run tests.
-    Converts file path to list of file paths or
-    creates list of file paths if input is a directory.
+def ppu2mpp(ppu: int, units: Union[str, int]) -> float:
+    """Convert pixels per unit (ppu) to microns per pixel (mpp)
 
     Args:
-        img_input (str or pathlib.Path): file path to images.
-        output_path (str or pathlib.Path): output directory path.
-        file_types (str): file types to process using cli.
-        mode (str): wsi or tile mode.
-        sub_dirname (str): name of subdirectory to save output.
+        ppu (int):
+            Pixels per unit.
+        units (Uniont[str, int]):
+            Units of pixels per unit. Valid options are "cm",
+            "centimeter", "inch", 2 (inches), 3(cm).
 
     Returns:
-        files_all (list): list of file paths to process.
-        output_path (pathlib.Path): updated output path.
+        mpp (float):
+            Microns per pixel.
 
     """
-    file_types = string_to_tuple(in_str=file_types)
-
-    if isinstance(output_path, str):
-        output_path = pathlib.Path(output_path)
-
-    if not os.path.exists(img_input):
-        raise FileNotFoundError
-
-    files_all = [
-        img_input,
-    ]
-
-    if os.path.isdir(img_input):
-        files_all = grab_files_from_dir(input_path=img_input, file_types=file_types)
-
-    if output_path is None and mode == "save":
-        input_dir = pathlib.Path(img_input).parent
-        output_path = input_dir / sub_dirname
-
-    if mode == "save":
-        output_path.mkdir(parents=True, exist_ok=True)
-
-    return files_all, output_path
-
-
-def prepare_model_cli(img_input, output_path, masks, file_types, mode):
-    """Prepares cli for running models.
-
-    Checks for existing directories to run tests.
-    Converts file path to list of file paths or
-    creates list of file paths if input is a directory.
-
-    Args:
-        img_input (str or pathlib.Path): file path to images.
-        output_path (str or pathlib.Path): output directory path.
-        masks (str or pathlib.Path): file path to masks.
-        file_types (str): file types to process using cli.
-        mode (str): wsi or tile mode.
-
-    Returns:
-        files_all (list): list of file paths to process.
-        masks_all (list): list of masks corresponding to input files.
-        output_path (pathlib.Path): output path
-
-    """
-    output_path = pathlib.Path(output_path)
-    file_types = string_to_tuple(in_str=file_types)
-
-    if output_path.exists():
-        raise FileExistsError("Path already exists.")
-
-    if not os.path.exists(img_input):
-        raise FileNotFoundError
-
-    if mode not in ["wsi", "tile"]:
-        raise ValueError("Please select wsi or tile mode.")
-
-    files_all = [
-        img_input,
-    ]
-
-    if masks is None:
-        masks_all = None
-    else:
-        masks_all = [
-            masks,
-        ]
-
-    if os.path.isdir(img_input):
-        files_all = grab_files_from_dir(input_path=img_input, file_types=file_types)
-
-    if os.path.isdir(str(masks)):
-        masks_all = grab_files_from_dir(input_path=masks, file_types=("*.jpg", "*.png"))
-
-    return files_all, masks_all, output_path
+    microns_per_unit = {
+        "centimeter": 1e4,  # 10,000
+        "cm": 1e4,  # 10,000
+        "mm": 1e3,  # 1,000
+        "inch": 25400,
+        "in": 25400,
+        2: 25400,  # inches in TIFF tags
+        3: 1e4,  # cm in TIFF tags
+    }
+    if units not in microns_per_unit:
+        raise ValueError(f"Invalid units: {units}")
+    return 1 / ppu * microns_per_unit[units]
